+++ week4/uchan0/mission/src/components/Header.jsx
@@ -1,0 +1,26 @@
+import React from "react";
+import ImageCollection from "../ImageCollection";
+
+const Header = () => {
+  return (
+    <header className="main-header">
+      <button type="button" className="location__container">
+        <div className="location__container--text">군자동</div>
+        <ImageCollection imgName={"down-chevron.svg"} />
+      </button>
+      <div className="header__icon-container">
+        <button type="button">
+          <ImageCollection imgName={"search.svg"} />
+        </button>
+        <button type="button">
+          <ImageCollection imgName={"hamburger.svg"} />
+        </button>
+        <button type="button">
+          <ImageCollection imgName={"bell.svg"} />
+        </button>
+      </div>
+    </header>
+  );
+};
+
+export default Header;
